+++ src/providers/emulatorjs/emulator-instance.ts
@@ -45,13 +45,14 @@ export type EjsInstance = EmulatorDiscInstance & EmulatorNativeSettingsInstance 
   takeScreenshot?: (source: string, format: string, upscale: number) => Promise<{blob?: Blob; screenshot?: unknown; format: string}>;
   downloadType?: {rom?: {dontExtractIfCore?: string[]}};
   on?: (event: string, callback: (...args: unknown[]) => void) => void;
+  callEvent?: (event: string, data?: unknown) => number;
 };
 
 export type EjsWindow = Window & {
   EJS_player?: string;
   EJS_core?: string;
   EJS_controlScheme?: string;
-  EJS_gameUrl?: string;
+  EJS_gameUrl?: string | File;
   EJS_gameName?: string;
   EJS_gameID?: number;
   EJS_pathtodata?: string;